chore: pages list layout consistency
