+++ web/components/pages/list/index.ts
@@ -1,6 +1,7 @@
 export * from "./applied-filters";
 export * from "./filters";
 export * from "./block";
+export * from "./block-item-action";
 export * from "./order-by";
 export * from "./root";
 export * from "./search-input";
